# audiocvt: Speed up channel count conversion by replacing divisions