Division is expensive and the compiler can not replace it themselves.
Though we now we will be ok in terms of floating point edge behaviour.

## Description

Replaces divisions by constants with reciprocal multiplication.
For 5.1 to stereo conversion this results in a 30% speedup.

**Edit** (revised statements about SSE/AVX, had a bug in test code):

I also vectorized the 51toStereo code which increased speed further.
~~I will test that code and remove draft status once I verified it.~~

**Edit**

conversion of `512 * 6` samples:
```
2021-06-12T02:00:19+02:00
Running cmake-build-relwithdebinfo/BenchVector51toStereo
Run on (8 X 4500 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.62, 0.72, 0.35
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
SDL_Convert51ToStereo               171 ns          171 ns      4095683
SDL_Convert51ToStereo_divfix        119 ns          119 ns      5949738
SDL_Convert51ToStereo_sse          50.3 ns         50.3 ns     13870364
SDL_Convert51ToStereo_avx          43.0 ns         43.0 ns     16289025
MEMCPY                             96.2 ns         96.2 ns      7287207
```

`SDL_Convert51ToStereo` is whats in `main`
`SDL_Convert51ToStereo_divfix` the first commit in this PR
`SDL_Convert51ToStereo_sse`  128bit 
`SDL_Convert51ToStereo_avx`  256bit