@flibitijibibo 2021-06-10
This is great! It's also pretty important for GetAudioDeviceSpec users since, [depending on how it's used](https://github.com/FNA-XNA/FAudio/blob/4881db35a409bc8cafa774440f6e672425853bcc/src/FAudio_platform_sdl2.c#L288-L331), downmixing will be really common. I'd be okay with having this patch in now, with the SSE/AVX work to come in later when it's ready.