@flibitijibibo 2021-06-10
I'm not 100% sure when 2.0.16 will freeze but looking at the milestone there should still be a good amount of time.

CC @icculus, this patch by itself should be pretty safe to merge.