@JoelLinn 2021-06-12
SSE saves 50% execution time compared to the C code (with the divisions replaced) while AVX only saves about 15% time compared to SSE.
This is because extracting the interleaved samples requires a lot of shuffles. AVX 256 bit registers are split in two lanes and only a select few instructions can move data between lo and hi.
AVX512F would help significantly since it can shuffle/permute values across all lanes at latency=1.

I noticed an Issue about the SSE3 version of Stereo to Mono. `__SSE3__` is not defined on Windows by default. If it were, it would mean the compiler is emitting SSE3 instructions anyways -  HaveSSE3() would be redundant. SDL would crash anyways in some random place were the compiler used an SSE3 instruction (On a system without SSE3). Imho, the correct approach is to not test for `__SSE3__` but for the availability of the intrinsics (like with `immintrin.h`)

Also, It would probably make sense to wrap the pragmas needed with GCC and Clang into a define in the `SDL_cpuinfo.h` header (?) and deploy them to other places were intrinsics are used (e.g. `SDL_ConvertStereoToMono`). I refrained from doing so myself because I am certain I would at least break one of the build system/platform combinations :P