@icculus 2021-07-24
> Division is expensive and the compiler can not replace it themselves.

I was going to say this is ridiculous, obviously a compiler can convert a division against a constant to a multiplication of its reciprocal, but oh my gosh, you're right:

https://gcc.godbolt.org/z/af19Kv43M

(Clang and GCC both won't do it, even at -O3)

...do they refuse to do this because of floating point precision concerns?