@icculus 2021-07-24
This all looks good, thanks!

I'm going to generate a couple separate issues for other comments, so we remember to explore them later.