@sezero 2021-07-24
The AVX patch is wrong and results in build failure, at least for x86:
```
src/audio/SDL_audiocvt.c: In function 'SDL_Convert51ToStereo_AVX':
src/audio/SDL_audiocvt.c:123: error: expected '=', ',', ';', 'asm' or '__attribute__' before 'two_fifths_v'
src/audio/SDL_audiocvt.c:123: error: 'two_fifths_v' undeclared (first use in this function)
src/audio/SDL_audiocvt.c:123: error: (Each undeclared identifier is reported only once
src/audio/SDL_audiocvt.c:123: error: for each function it appears in.)
src/audio/SDL_audiocvt.c:123: warning: implicit declaration of function '_mm256_set1_ps'
src/audio/SDL_audiocvt.c:124: error: expected '=', ',', ';', 'asm' or '__attribute__' before 'half'
src/audio/SDL_audiocvt.c:124: error: ISO C90 forbids mixed declarations and code
src/audio/SDL_audiocvt.c:124: error: 'half' undeclared (first use in this function)
src/audio/SDL_audiocvt.c:146: error: '__m256' undeclared (first use in this function)
src/audio/SDL_audiocvt.c:146: error: expected ';' before 'fc_distributed'
src/audio/SDL_audiocvt.c:149: error: ISO C90 forbids mixed declarations and code
src/audio/SDL_audiocvt.c:153: error: expected ';' before 'permuted0'
src/audio/SDL_audiocvt.c:155: error: expected ';' before 'permuted1'
src/audio/SDL_audiocvt.c:160: error: expected ';' before 'out'
src/audio/SDL_audiocvt.c:161: error: 'out' undeclared (first use in this function)
src/audio/SDL_audiocvt.c:161: warning: implicit declaration of function '_mm256_add_ps'
src/audio/SDL_audiocvt.c:161: error: 'fc_distributed' undeclared (first use in this function)
src/audio/SDL_audiocvt.c:162: warning: implicit declaration of function '_mm256_mul_ps'
src/audio/SDL_audiocvt.c:164: warning: implicit declaration of function '_mm256_storeu_ps'
src/audio/SDL_audiocvt.c:151: warning: unused variable 'permuted0_hi'
src/audio/SDL_audiocvt.c:149: warning: unused variable 'permuted0_lo'
src/audio/SDL_audiocvt.c:144: warning: unused variable 'fc_distributed_hi'
src/audio/SDL_audiocvt.c:142: warning: unused variable 'fc_distributed_lo'
src/audio/SDL_audiocvt.c:138: warning: unused variable 'in4'
src/audio/SDL_audiocvt.c:135: warning: unused variable 'in1'
make: *** [build/SDL_audiocvt.lo] Error 1
```