@JoelLinn 2021-07-24
AVX should be available to 32 bit mode but it seems like either the header isn't included or they are undefined for another reason