@sezero 2021-07-24
Hm, old compilers have issues:

- gcc-4.4.7 (linux): fails as above

- gcc-4.5.4 (mingw): ICEs with
 `src/audio/SDL_audiocvt.c:113:9: internal compiler error: in c_builtin_function_ext_scope, at c-decl.c:3558`
  Line 113 is `#pragma GCC target("avx")`

- gcc-4.9.4 (linux): no failures.