@Triang3l 2021-07-24
> ...do they refuse to do this because of floating point precision concerns?
Optimization that may affect rounding (including replacement of division with multiplication, reordering according to associativity properties, usage of fused multiply-add on PowerPC, among other things) can only be performed with the `-ffast-math` (aka `/fp:fast` on Visual C++) compiler option. Also, while 2.5 is exactly 0b1.01 * 2^1, 0.4 cannot even be represented as a binary fraction — in binary that's 0b1.100110011… * 2^-2, so if you replace the division by a multiplication in this case, you will be multiplying by an already rounded factor. The only case when replacement of division with multiplication is exact, and is allowed with the precise or the strict math option, is division by powers of 2, in which case both are simple exponent biasing.