@sezero 2021-07-24
>     * gcc-4.5.4 (mingw): ICEs with
>       `src/audio/SDL_audiocvt.c:113:9: internal compiler error: in c_builtin_function_ext_scope, at c-decl.c:3558`
>       Line 113 is `#pragma GCC target("avx")`

BTW, that ICE happens when targeting x86_64 too, so a gcc problem, to be ignored.