@JoelLinn 2021-07-24
"Function Specific Option Pragmas" have been introduced in GCC 4.4, so every earlier compiler should thus crash.
https://gcc.gnu.org/onlinedocs/gcc-4.4.7/gcc/Function-Specific-Option-Pragmas.html

If 4.5.4 crashes (ignoring it's mingw) I wouldn't expect 4.4.x to work either.
Maybe we should just disable all the affected code for such old compilers, before the pragma is parsed (also concerns #4531)