@icculus 2021-07-24
> Maybe we should just disable all the affected code for such old compilers, before the pragma is parsed

Yes, this needs to be disabled on older compilers if they will ICE on it. We can just make a blanket "if it's less than GCC 5.0.0" or whatever. Those compilers will just not get the SSE version of the converter.