@icculus 2021-07-24
Can't reopen pull requests, but I opened a new issue for the compiler ICE'ing problem.